ECPG: Start tracking subtransactions and cursors in the runtime library to reduce network turnaround.
For this purpose, the arguments to ECPGtrans() have changed and
4 new public functions were introduced: ECPGopen(), ECPGfetch(),
ECPGcursor_dml() and ECPGclose(). The previously exposed ecpg_do()
is used to drive all these functions. The library version was
changed to 7.0 from 6.6 because of the ECPGtrans change.

New error handling is added to mimic the server behaviour if a
previous function checking the savepoint or cursor name failed.

A new regression test was added to exercise certain aspects of
the client-side tracking of cursors and subtransactions.